getting-started: Bring toolbox up as peer, rework rpm-ostree
I think `toolbox` should be a clear peer of flatpak and rpm-ostree.
At some point I'd like to rework this page so we move `toolbox`
basics here, and demote `rpm-ostree` details to a separate page.

The rpm-ostree section is reworked to give clear examples of
packages to try layering and that need to be layered if desired.
And similarly give examples for `override replace`.  Let's not
emphasize `override remove` at all as its uses are obscure.