Add a section on resetting toolbox permissions
A while back I ran into an error with toolbox where I could not enter it due to faulty permissions. This may be fixed in newer versions of libpod already, I have not had the same error since the first time, but just in case someone encounters a similar issue this may help them resolve the issue. Credit goes to refi64 on discussion.fedoraproject.org for suggesting this solution that fixed the problem when I encountered it.