Create new section in docs
I wanted to review issue #70 with kargs options. As I was gathering information about it, I came up with an idea for a new section in documentation, rpm-ostree. I think it would be nice starting point for a total newcomers to Silverblue and hybrid packaging systems in general. I tried to make it look like toolbox section in docs. It should provide some surface level knowledge of rpm-ostree.