Update toolbox.adoc (#83)
Buildah is no longer used in Toolbox to (re)create images.

`toolbox run` is also a command of Toolbox that needs to be mentioned.

Some users that use `zsh` with `powerline` or something similar sometimes get confused by the lack of the diamond symbol that signalizes being in a toolbox. Before the profile file is updated, it should be documented.

The 'How it works' section did not mention `flatpak-spawn` which is used to execute commands on the host.