toolbox: Adjust style
In my previous commit I made an oversight and one of the blocks was not
properly styled.

Also, until now the "Commands and usage" section used question/answers
format for command description. I expanded the description in the
previous commit and that caused to show the shortcomming of the format.
Only the first paragraph is considered the answer. This causes
misallignment of the text. With this the questions are turned into 3rd
level titles. This also allows to link to the command description.