updates: Clarify that we do not support skipping major releases
We do not test or support skipping major versions of Fedora for
rpm-ostree variants unlike classic DNF based ones.

This is due to the fact that changes might require a recent version of
rpm-ostree to be able to rebase to the next version. One example is the
bdb -> sqlite change.